backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
supabase==2.5.0
reportlab==4.2.2
yt-dlp==2024.8.6
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
scikit-learn==1.5.1
requests==2.32.3
